Tests/LibWeb/Text/expected/HTML/HTMLMediaElement-adopted-from-inactive-document.txt

video loadstart event (networkState=2, error=null)
video error event (networkState=3, error=4)
window load event
